requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
click==8.1.3
colorama==0.4.6
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
importlib-metadata==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==2.1.2
PyMySQL==1.0.2
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
zipp==3.15.0
